refactorización del proyecto terminada
