+++ components/common/componentsCrearLicitacion/InfoDetallada.tsx
@@ -1,8 +1,9 @@
 import { Dispatch, SetStateAction } from 'react'
+import { OPTIONS_ESTADO } from '../../../consts'
 import { FormCrearLicitacionUser, HandlerChange } from '../../../types/form'
-import { Estado } from '../../../types/form/enums'
 import { ErrorsForm } from '../../../types/validators'
 import generateNumber from '../../../utils/generateNumber.utility'
+import InputSelect from '../inputs/InputSelect'
 
 type Props={
     handleChange:HandlerChange,
@@ -28,15 +29,7 @@ export default function InfoDetallada ({ handleChange, step, setStep, form, setF
                             <label className="text-gray-500 text-sm" htmlFor="requisitos">Requisitos</label>
                             <textarea onChange={handleChange} value={form.requisitos} name="requisitos" className="rounded dark:bg-gray-800 placeholder:text-sm" placeholder="Agregar requisitos de licitación" id="" cols={30} rows={5}></textarea>
                         </article>
-                        <article className="flex flex-col my-4">
-                            <label className="text-gray-500 text-sm" htmlFor="estado">Estado <strong className='text-red-500' >*</strong></label>
-                            <select onChange={handleChange} value={form.estado} className="dark:bg-gray-800 dark:text-gray-400" name="estado" id="">
-                                <option value="">-Seleccione un estado-</option>
-                                <option value={Estado.Cerrado}>Cerrado</option>
-                                <option value={Estado.Abierto}>Abierto</option>
-                            </select>
-                            {error.estado && <p className='text-red-500 font-light text-sm' >{error.estado}</p> }
-                        </article>
+                        <InputSelect error={error.estado} handleChange={handleChange} label="Estado" name='estado' options={OPTIONS_ESTADO} value={form.estado} />
                         <article className="flex flex-col my-4">
                             <label className="text-gray-500 dark:text-gray-400 text-sm" htmlFor="author">Responsable de Licitación <strong className='text-red-500' >*</strong></label>
                             <input onChange={handleChange} value={form.author} name="author" className="rounded dark:bg-gray-800 dark:text-gray-400 placeholder:text-sm " placeholder="Nombre completo" type="text" />
