+++ components/common/componentsCrearLicitacion/InfoGeneral.tsx
@@ -3,6 +3,7 @@ import { AiOutlineQuestionCircle } from 'react-icons/ai'
 import { DataSelect } from '../../../types/models'
 import { FormCrearLicitacionUser, HandlerChange } from '../../../types/form'
 import { ErrorsForm } from '../../../types/validators'
+import InputSelect from '../inputs/InputSelect'
 type Props={
     step:number,
     setStep:Dispatch<SetStateAction<number>>,
@@ -48,16 +49,7 @@ export default function InfoGeneral ({ step, setStep, handleChange, form, servic
                                 <option value="Tipo03">Licitación Tipo 03</option>
                             </select>
                         </article>
-                        <article className="flex flex-col my-4">
-                            <label className="text-gray-500 dark:text-gray-400 text-sm 2xl:text-lg " htmlFor="tipoServicio">Tipo de Servicio <strong className='text-red-500' >*</strong></label>
-                            <select onChange={handleChange} value={form.tipoServicio} className="dark:bg-gray-800 dark:text-gray-400" name="tipoServicio" id="">
-                                <option value="">-Seleccionar un tipo de Servicio</option>
-                                {servicios.map((el) => (
-                                    <option key={el._id} value={el._id}>{el.name}</option>
-                                ))}
-                            </select>
-                            {error.tipoServicio && <p className='text-red-500 font-light text-sm' >{error.tipoServicio}</p> }
-                        </article>
+                        <InputSelect error={error.tipoServicio} handleChange={handleChange} label="Tipo de Servicio" name='tipoServicio' options={servicios.map((item) => ({ label: item.name, value: item.name }))} value={form.tipoServicio} />
                         <article className="flex justify-end pt-4">
                             <span onClick={() => setStep(step + 1)} className="bg-green-600 py-2 px-4 text-white block cursor-pointer">Continuar</span>
                         </article>
