+++ consts/calendar.ts
@@ -1,0 +1,2 @@
+export const CALENDAR_SIZE = [...Array(42).keys()]
+export const DAYS_WEEK = ['d', 'l', 'm', 'm', 'j', 'v', 's']
