+++ consts/heads.ts
@@ -1,0 +1,1 @@
+export const HEADS_TABLE_ESPECIFICACION_MES = ['mes', 'hp (mw)', 'hfp (mw)']
