+++ types/form/Select.d.ts
@@ -1,0 +1,4 @@
+export type OptionSelect={
+    value:string,
+    label:string
+}
