+++ types/form/enums.ts
@@ -4,3 +4,8 @@ export enum Estado{
     Proceso='en proceso',
     Cerrado='cerrado'
 }
+export enum PotenciaFacturar{
+    MD_mes='MD en Horas Puntas del Mes',
+    MD_personalizada='Demanda coincidente con la Máxima Demanda del SEIN',
+    SEIN='MD en Horas de Punta personalizada'
+}
