+++ types/form/index.d.ts
@@ -5,3 +5,4 @@ export * from './FormCrearLicitacionUser'
 export * from './FormCrearOfertaProveedor'
 export * from './NumMes'
 export * from './Handlers'
+export * from './Select'
