+++ types/hooks/useCalendar.d.ts
@@ -1,0 +1,13 @@
+export type UseCalendar=(fechaInicioApertura:string, fechaFinApertura:string)=>{
+    both:boolean,
+    cal:{
+        fechaInicio:{
+            day:number,
+            dayWeek:number
+        },
+        fechaFin:{
+            day:number,
+            dayWeek:number
+        }
+    }
+}
