+++ types/hooks/useData.d.ts
@@ -1,7 +1,7 @@
 import { Session } from 'next-auth'
 import { DataSelect } from '../models'
 
-export type UseData = (session: Session) => {
+export type UseData = (session: Session|null) => {
     brgs:Array<DataSelect>,
     puntoSums:Array<DataSelect>,
     servicios:Array<DataSelect>
