+++ types/hooks/useMeses.d.ts
@@ -1,0 +1,4 @@
+import { Dispatch, SetStateAction } from 'react'
+import { FormCrearLicitacionUser } from '../form'
+
+export type UseMeses = (update:boolean, form:FormCrearLicitacionUser, setForm:Dispatch<SetStateAction<FormCrearLicitacionUser>>) => void
